requirements.txt
customtkinter==5.2.1
PyMuPDF==1.23.8
Pillow==10.2.0
pyinstaller==6.3.0
qrcode==7.4.2
